[#618] Fix payload_len handling
Use size_t/ssize_t to store the message buffer sizes instead of
a uint16_t and int types.

+++ core/packet.c
@@ -386,7 +386,7 @@ static int prv_change_to_block1(lwm2m_context_t * contextP, void * sessionH, uin
     
     transaction = prv_get_transaction(contextP, sessionH, mid);
 
-    for (int n = 1; 16 << n <= (int)size ; n++) {
+    for (uint16_t n = 1; 16 << n <= (uint16_t)size; n++) {
         block_size = 16 << n;
     }
 
@@ -473,16 +473,15 @@ static int prv_send_get_next_block2(lwm2m_context_t * contextP,
  * Erbium is Copyright (c) 2013, Institute for Pervasive Computing, ETH Zurich
  * All rights reserved.
  */
-void lwm2m_handle_packet(lwm2m_context_t * contextP,
-                         uint8_t * buffer,
-                         int length,
-                         void * fromSessionH)
-{
+void lwm2m_handle_packet(lwm2m_context_t *contextP, uint8_t *buffer, size_t length, void *fromSessionH) {
     uint8_t coap_error_code = NO_ERROR;
     static coap_packet_t message[1];
     static coap_packet_t response[1];
 
     LOG("Entering");
+    /* The buffer length is uint16_t here, as UDP packet length field is 16 bit.
+     * This might change in the future e.g. for supporting TCP or other transport.
+     */
     coap_error_code = coap_parse_message(message, buffer, (uint16_t)length);
     if (coap_error_code == NO_ERROR)
     {
@@ -854,7 +853,6 @@ void lwm2m_handle_packet(lwm2m_context_t * contextP,
         message_send(contextP, message, fromSessionH);
     }
 }
-
 
 uint8_t message_send(lwm2m_context_t * contextP,
                      coap_packet_t * message,
